+++ src/main/kotlin/uk/co/ben_gibson/git/link/ui/actions/gutter/BrowserAction.kt
@@ -1,0 +1,3 @@
+package uk.co.ben_gibson.git.link.ui.actions.gutter
+
+class BrowserAction : GutterAction(Type.BROWSER)