+++ src/main/kotlin/uk/co/ben_gibson/git/link/ui/actions/gutter/CopyAction.kt
@@ -1,0 +1,3 @@
+package uk.co.ben_gibson.git.link.ui.actions.gutter
+
+class CopyAction : GutterAction(Type.COPY)